+++ src/acl/acl.ts
@@ -294,18 +294,16 @@ export function createAclFromFallbackAcl(
     fallbackAclRules,
     resource.internal_acl.fallbackAcl.internal_accessTo
   );
-  const resourceAclRules = defaultAclRules.map((rule) => {
+  const newAclRules = defaultAclRules.map((rule) => {
     rule = removeAll(rule, acl.default);
     rule = removeAll(rule, acl.defaultForNew);
     rule = setIri(rule, acl.accessTo, getSourceUrl(resource));
+    rule = setIri(rule, acl.default, getSourceUrl(resource));
     return rule;
   });
 
   // Iterate over every ACL Rule we want to import, inserting them into `emptyResourceAcl` one by one:
-  const initialisedResourceAcl = resourceAclRules.reduce(
-    setThing,
-    emptyResourceAcl
-  );
+  const initialisedResourceAcl = newAclRules.reduce(setThing, emptyResourceAcl);
 
   return initialisedResourceAcl;
 }